requirements.txt
albumentations==0.5.2
pandas
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.2.9
opencv-python>=4.2.0.34
Pillow>=6.1.0
pretrainedmodels==0.7.4
scikit-image>=0.15.0
scikit-learn==0.21.2
timm
torch_optimizer
torch>=1.4.0
torchvision>=0.5.0
tqdm
wandb